move identity status check to background
